Backend/requirements.txt
fastapi>=0.95.0
uvicorn[standard]>=0.20.0
google-cloud-storage>=2.10.0
google-cloud-secret-manager>=2.16.0
google-auth>=2.20.0
google-generativeai>=0.3.0
anthropic>=0.8.0
python-dotenv>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0